Remove MonitorThread
