+++ src/gcmon/__init__.py
@@ -3,7 +3,6 @@ __version__ = "0.1.0"
 from .child_process_runner import ChildProcess, ChildProcessRunner
 from .exporters import EventsExporter, JsonlExporter, StdoutExporter, TraceExporter
 from .monitor import EventsMonitor, create_monitor
-from .monitor_thread import MonitorThread
 
 __all__ = [
     "ChildProcess",
@@ -11,7 +10,6 @@ __all__ = [
     "EventsExporter",
     "EventsMonitor",
     "JsonlExporter",
-    "MonitorThread",
     "StdoutExporter",
     "TraceExporter",
     "__version__",
